requirements.txt
flask==3.0.3
httpx==0.27.2
azure-identity==1.17.1
azure-data-tables==12.6.0
azure-keyvault-secrets==4.8.0
openai==1.42.0
python-dotenv==1.0.1